Move single function out of file so file can be removed
